+++ conf/db/migrations/20240303082633_user_groups.php
@@ -19,13 +19,15 @@ final class UserGroups extends AbstractMigration
      */
     public function change(): void
     {
-        $this->table('group')
+        $this->table('group', ['id' => false, 'primary_key' => ['id']])
+            ->addColumn('id', 'integer', ['identity' => true, 'signed' => false]) // UNSIGNED AUTO_INCREMENT PK
             ->addColumn('created', 'timestamp', ['default' => 'CURRENT_TIMESTAMP'])
             ->addColumn('name_public', 'string', ['limit' => 255])
             ->addColumn('internal_notes', 'text', ['null' => true])
             ->create();
 
-        $this->table('user_group')
+        $this->table('user_group', ['id' => false, 'primary_key' => ['id']])
+            ->addColumn('id', 'integer', ['identity' => true, 'signed' => false]) // UNSIGNED AUTO_INCREMENT PK
             ->addColumn('created', 'timestamp', ['default' => 'CURRENT_TIMESTAMP'])
             ->addColumn('user_id', 'integer', ['signed' => false])
             ->addForeignKey('user_id', 'users', 'id', ['delete' => 'CASCADE', 'update' => 'NO_ACTION'])
@@ -34,7 +36,8 @@ final class UserGroups extends AbstractMigration
             ->addIndex(['user_id'])
             ->create();
 
-        $this->table('group_activation_tokens')
+        $this->table('group_activation_tokens', ['id' => false, 'primary_key' => ['id']])
+            ->addColumn('id', 'integer', ['identity' => true, 'signed' => false]) // UNSIGNED AUTO_INCREMENT PK
             ->addColumn('created', 'timestamp', ['default' => 'CURRENT_TIMESTAMP'])
             ->addColumn('group_id', 'integer', ['signed' => false])
             ->addForeignKey('group_id', 'group', 'id', ['delete' => 'CASCADE', 'update' => 'NO_ACTION'])