+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Added` for new features
 
-- FACEBOOK_API_VERSION is 'v21.0' by default, but may be configured by the app
-- add a flag 'AuthApp:FLAG_USER_ROUTE_NOT_USED_FOR_SOCIAL_LOGIN') to trigger only reload, while the default behaviour is to go to the USER_ROUTE after a successful social login
+- AuthConstant::FACEBOOK_API_VERSION is 'v21.0' by default, but may be configured by the app
+- if AuthConstant::SOCIAL_LOGIN_SUCCESS_URL is set and not empty, then there will be a social login success page
 
 ### `Changed` for changes in existing functionality